+++ djangosettings.py
@@ -489,6 +489,17 @@ EMAIL_HOST = None
 EMAIL_PORT = 25
 EMAIL_BACKEND = "django.core.mail.backends.smtp.EmailBackend"
 
+# Configuration of the ftp/sftp/ftps server where to upload reports
+# Note that for SFTP protocol, the host needs to be defined
+# in the known_hosts file
+FTP_PROTOCOL = "SFTP"  # supported protocols are SFTP, FTPS and FTP (unsecure)
+FTP_HOST = "ftp.server.com"
+FTP_PORT = 22
+FTP_USER = "user"
+FTP_PASSWORD = "password"
+FTP_FOLDER = "/tmp"  # folder where the files should be uploaded
+
+
 # Port number when not using Apache
 PORT = 8000