+++ doc/release-notes.rst
@@ -17,6 +17,11 @@ Release notes
 
 - | Bug fix: formatting of negative numbers was showing too many digits after the decimal.
 
+.. rubric:: Integration
+
+- | Addition of the **upload exported reports** menu in the execute screen, allowing users to export selected
+    reports to a third-party server (using a ftp, sftp or ftps connection).
+
 6.13.0 (2021/04/20)
 ===================